Add files via upload
Setup a back-end server for the previous application.
This needs to be a separate application. Create CRUD
APIs (REST) for user model. Store data in JSON file.
Use that as a source of data. UI functionality should
remain the same. Everything should be in TS. Use express
for REST APIs. Push the code to a separate repo on
Github. Generate a PR for review.